+++ model_navigator/kubernetes/evaluator/deployment.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2021-2022, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -38,9 +38,9 @@ class Deployment(helm.Deployment):
             },
         }
 
-        env = list()
-        volumeMounts = list()
-        volumeAttach = list()
+        env = []
+        volumeMounts = []
+        volumeAttach = []
 
         self._prepare_volumes(env, volumeAttach, volumeMounts)