+++ model_navigator/model_analyzer/profiler.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2021-2022, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -175,7 +175,7 @@ class ProfileConfigGenerator(BaseConfigGenerator):
             for model_name in models_list
         )
 
-        # https://github.com/triton-inference-server/model_analyzer/blob/r21.12/docs/config.md
+        # https://github.com/triton-inference-server/model_analyzer/blob/r22.01/docs/config.md
         config = {
             "run_config_search_disable": manual_config_search,
             "profile_models": models_list,